populate_db/requirements.txt
beautifulsoup4
biopython
bokeh==2.3.2
click
fastcluster
git+https://github.com/Maayanlab/maayanlab-bioinformatics.git
html5lib
jpype1
lxml
nbformat
networkx
nltk
numpy
odfpy
openpyxl
pandas
plotly
psycopg2-binary
python-docx
python-dotenv
pyxlsb
requests
s3fs
seaborn
tabula-py
tqdm
umap-learn
xlrd
GEOparse